requirements.txt
paramiko
boto3
jsonschema
pyyaml
